server/requirements.txt
python-chess
sqlalchemy
flask
pyserial
flask-cors
flask-socketio
eventlet  # Optional but recommended for better WebSocket performance
aiohttp   # Needed by socketio.AsyncClient
